+++ material/src/directives/types.ts
@@ -1,0 +1,1 @@
+export * from './dialogPopup/dialogPopup.directive';