+++ material/src/material.ts
@@ -3,3 +3,4 @@ export * from './components/basicDialogPopup/types';
 export * from './plugins/popup/components';
 export * from './plugins/popup';
 export * from './components/basicDialogPopup';
+export * from './directives/types';